moved core code inside core module
