+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/StageExecutionManager.java
@@ -1,17 +1,17 @@
-package com.github.nishgpt.chainexecutor;
+package com.github.nishgpt.chainexecutor.core;
 
-import com.github.nishgpt.chainexecutor.exceptions.ChainExecutorException;
-import com.github.nishgpt.chainexecutor.exceptions.ErrorCode;
-import com.github.nishgpt.chainexecutor.models.execution.ExecutionContext;
-import com.github.nishgpt.chainexecutor.models.execution.ExecutorAuxiliaryKey;
-import com.github.nishgpt.chainexecutor.models.execution.StageExecutionRequest;
-import com.github.nishgpt.chainexecutor.models.execution.StageExecutor;
-import com.github.nishgpt.chainexecutor.models.execution.StageExecutorFactory;
-import com.github.nishgpt.chainexecutor.models.execution.StageExecutorKey;
-import com.github.nishgpt.chainexecutor.models.stage.Stage;
-import com.github.nishgpt.chainexecutor.models.stage.StageChainIdentifier;
-import com.github.nishgpt.chainexecutor.models.stage.StageChainRegistry;
-import com.github.nishgpt.chainexecutor.models.stage.StageStatus;
+import com.github.nishgpt.chainexecutor.core.exceptions.ChainExecutorException;
+import com.github.nishgpt.chainexecutor.core.exceptions.ErrorCode;
+import com.github.nishgpt.chainexecutor.core.models.execution.ExecutionContext;
+import com.github.nishgpt.chainexecutor.core.models.execution.ExecutorAuxiliaryKey;
+import com.github.nishgpt.chainexecutor.core.models.execution.StageExecutionRequest;
+import com.github.nishgpt.chainexecutor.core.models.execution.StageExecutor;
+import com.github.nishgpt.chainexecutor.core.models.execution.StageExecutorFactory;
+import com.github.nishgpt.chainexecutor.core.models.execution.StageExecutorKey;
+import com.github.nishgpt.chainexecutor.core.models.stage.Stage;
+import com.github.nishgpt.chainexecutor.core.models.stage.StageChainIdentifier;
+import com.github.nishgpt.chainexecutor.core.models.stage.StageChainRegistry;
+import com.github.nishgpt.chainexecutor.core.models.stage.StageStatus;
 import java.util.Objects;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
