+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/models/execution/ExecutionContext.java
@@ -1,4 +1,4 @@
-package com.github.nishgpt.chainexecutor.models.execution;
+package com.github.nishgpt.chainexecutor.core.models.execution;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
