+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/models/execution/StageExecutionRequest.java
@@ -1,0 +1,8 @@
+package com.github.nishgpt.chainexecutor.core.models.execution;
+
+import com.github.nishgpt.chainexecutor.core.models.stage.Stage;
+
+public interface StageExecutionRequest {
+
+  Stage getStage();
+}
