+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/models/execution/StageExecutorFactory.java
@@ -1,4 +1,4 @@
-package com.github.nishgpt.chainexecutor.models.execution;
+package com.github.nishgpt.chainexecutor.core.models.execution;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
