+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/models/execution/StageExecutorKey.java
@@ -1,6 +1,6 @@
-package com.github.nishgpt.chainexecutor.models.execution;
+package com.github.nishgpt.chainexecutor.core.models.execution;
 
-import com.github.nishgpt.chainexecutor.models.stage.Stage;
+import com.github.nishgpt.chainexecutor.core.models.stage.Stage;
 import lombok.Builder;
 import lombok.Data;
