+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/models/stage/Stage.java
@@ -1,0 +1,6 @@
+package com.github.nishgpt.chainexecutor.core.models.stage;
+
+public interface Stage {
+
+  String name();
+}
