+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/models/stage/StageChainIdentifier.java
@@ -1,0 +1,5 @@
+package com.github.nishgpt.chainexecutor.core.models.stage;
+
+public interface StageChainIdentifier {
+
+}
