--- src/main/java/com/github/nishgpt/chainexecutor/models/execution/StageExecutionRequest.java
@@ -1,8 +1,0 @@
-package com.github.nishgpt.chainexecutor.models.execution;
-
-import com.github.nishgpt.chainexecutor.models.stage.Stage;
-
-public interface StageExecutionRequest {
-
-  Stage getStage();
-}
