--- src/main/java/com/github/nishgpt/chainexecutor/models/stage/Stage.java
@@ -1,6 +1,0 @@
-package com.github.nishgpt.chainexecutor.models.stage;
-
-public interface Stage {
-
-  String name();
-}
