--- src/main/java/com/github/nishgpt/chainexecutor/models/stage/StageChainIdentifier.java
@@ -1,5 +1,0 @@
-package com.github.nishgpt.chainexecutor.models.stage;
-
-public interface StageChainIdentifier {
-
-}
